+++ v2g-liberty/rootfs/root/appdaemon/apps/v2g_liberty/main_app.py
@@ -280,8 +280,7 @@ class V2Gliberty:
                 # Cancel previous scheduling timers as they might have discharging instructions
                 # as well.
                 self.__log(
-                    f"set_next_action | start Boost charge as soc {soc} is "
-                    f"below minimum '{c.CAR_MIN_SOC_IN_PERCENT}'."
+                    f"Start Boost charge: SoC '{soc}%' < minimum '{c.CAR_MIN_SOC_IN_PERCENT}%'."
                 )
                 self.__cancel_charging_timers()
                 await self.__start_max_charge_now()
@@ -581,13 +580,9 @@ class V2Gliberty:
             attempt = kwargs.get("attempt", 0) + 1
             if attempt == 5:
                 self.__log(
-                    "Failed to log_version after 5 attempts, aborting.", level="WARNING"
+                    "Failed to log_version to FM after 5 attempts, aborting.", level="WARNING"
                 )
                 return
-            self.__log(
-                f"Attempt {attempt} to log_version failed, retrying in 15 sec.",
-                level="WARNING",
-            )
             await self.hass.run_in(self.__log_version, delay=15, attempt=attempt)
 
     async def __handle_car_connect(self):
@@ -607,6 +602,9 @@ class V2Gliberty:
         """
         # Reset any possible target for discharge due to SoC > max-soc
         self.back_to_max_soc = None
+
+        # Just to be sure, this also is done when disconnect ir requested
+        self.in_boost_to_reach_min_soc = False
 
         # There might be a notification to ask the user to dismiss an event or not,
         # if the car gets disconnected this notification can be removed.
@@ -853,6 +851,12 @@ class V2Gliberty:
         old_state = old.get("state", None)
         self.__log(f"Charge mode has changed from '{old_state}' to '{new_state}'")
 
+        await self.__clear_all_soc_chart_lines()
+
+        if old_state == "Automatic":
+            self.__log("Cancel scheduled charging (timers).")
+            self.__cancel_charging_timers()
+
         if (
             old_state in ["Max boost now", "Max discharge now"]
             and new_state == "Automatic"
@@ -865,19 +869,12 @@ class V2Gliberty:
                     "source": "Reset for 'Max boost now' to 'Automatic'",
                 }
             )
-        if old_state == "Automatic":
-            self.__log(
-                "Charge mode changed from Automatic to Max boost now, "
-                "cancel timer and start charging."
-            )
-            self.__cancel_charging_timers()
 
         if old_state != "Stop" and new_state == "Stop":
             # New mode "Stop" is handled by set_next_action
             self.__log(
                 "Stop charging (if in action) and give control based on charge_mode = Stop"
             )
-            await self.__clear_all_soc_chart_lines()
             self.in_boost_to_reach_min_soc = False
             await self.evse_client_app.set_inactive()
             # For monitoring
@@ -919,6 +916,11 @@ class V2Gliberty:
         """
         self.__log("charger disconnect requested")
         await self.__reset_no_new_schedule()
+
+        # When disconnection during boost_to_reach_min_soc this needsa to be reset to allow
+        # reactivating this mode when car does not get disconnected.
+        # This also is set when car is actually disconnected
+        self.in_boost_to_reach_min_soc = False
 
         await self.evse_client_app.stop_charging()
         # Control is not given to user, this is only relevant if charge_mode is "Off" (stop).