+++ v2g-liberty/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- 🪲 BUG: Forced charge not always working (#337)
 - 🪲 BUG: Calendare reservations wrongly visualised in chart (#335)
 
 ### Added